confirm_delete and sort array

+++ src/App.jsx
@@ -1,12 +1,20 @@
-import { useState } from 'react';
+import { useState, useRef } from 'react';
 import './App.css';
 import Add from './pages/add-item/add-item.page'
 import View from './pages/view-item/view'
 import Header from './components/header/header.component';
-
+import Dialog from './components/Dialog/dialog'
 
 function App() {
-  const [items, setItems] = useState(JSON.parse(localStorage.getItem('todoList'))||'[]');
+
+
+  //You can put all product information into diaglog
+  const [dialog, setDialog] = useState({
+    message: "",
+    isLoading: false,
+  });
+
+  const [items, setItems] = useState(JSON.parse(localStorage.getItem('todoList')) || '[]');
   //const [Window,setWindow] =useState(false);
   const [currentpage, setCurrentPage] = useState(items.length > 0 ? 'view' : 'add');
 
@@ -20,13 +28,37 @@ function App() {
 
   }
 
-  const deleteItem = (id) => {
-   // setWindow(true);
-    const newitems = items.filter(item => item.id !== id);
-    setItems(newitems);
-    localStorage.setItem('todoList', JSON.stringify(newitems));
+  const idItemRef = useRef();
 
-  }
+  const handleDialog = (message, isLoading) => {
+    setDialog({ message, isLoading });
+  };
+
+  const deleteItem = (id) => {
+    // const index = items.findIndex((item) => item.id === id);
+    handleDialog("Are you sure you want to delete? sure you are sure ?", true);
+    idItemRef.current = id;
+  };
+
+
+  const areUSureDelete = (choose) => {
+    if (choose) {
+
+      const newitems = items.filter(item => item.id !== idItemRef.current);
+      setItems(newitems);
+      localStorage.setItem('todoList', JSON.stringify(newitems));
+      handleDialog("", false);
+    } else {
+      handleDialog("", false);
+    }
+  };
+
+  const sortFunction = (item) => {
+    const newItems = [...item].sort((firstItem, secondItem) =>
+      firstItem.isDone > secondItem.isDone ? 1 : -1,
+    );
+    setItems(newItems);
+  };
 
   const finishitem = (id) => {
     // const newitem = items.map((item) => {
@@ -39,6 +71,7 @@ function App() {
     // Sort version of the mapping
     const newItems = items.map(item => item.id === id ? { ...item, isDone: true } : item);
     setItems(newItems);
+    sortFunction(newItems);
     localStorage.setItem('todoList', JSON.stringify(newItems));
 
   }
@@ -47,7 +80,7 @@ function App() {
     <div className="App">
       <Header setCurrentPage={setCurrentPage} currentpage={currentpage} />
       <div className="container">
-      {currentpage === 'add' && <Add onAddItem={addItem} />}
+        {currentpage === 'add' && <Add onAddItem={addItem} />}
         {currentpage === 'view' &&
           <View
             items={items}
@@ -55,6 +88,12 @@ function App() {
             onFinish={finishitem} />
         }
       </div>
+      {dialog.isLoading && (
+        <Dialog
+          onDialog={areUSureDelete}
+          message={dialog.message}
+        />
+      )}
     </div>
 
   );
